requirements.txt
aiogram==3.23.0
aiohttp==3.13.2
python-dotenv==1.2.1